hard refactoring

+++ vellum/acpi/isr.c
@@ -9,7 +9,7 @@ uacpi_status uacpi_kernel_install_interrupt_handler(
 {
     status_t status;
     
-    status = _pc_isr_add_interrupt_handler(irq, ctx, (interrupt_handler_t)handler, (struct isr_handler **)out_irq_handle);
+    status = _pc_isr_add_interrupt_handler((int)irq, ctx, (interrupt_handler_t)handler, (struct isr_handler **)out_irq_handle);
     if (!CHECK_SUCCESS(status)) return UACPI_STATUS_INTERNAL_ERROR;
 
     return UACPI_STATUS_OK;
